+++ wear/src/main/java/net/osdn/gokigen/joggingtimer/recordlist/IDetailLauncher.java
@@ -2,6 +2,6 @@ package net.osdn.gokigen.joggingtimer.recordlist;
 
 public interface IDetailLauncher
 {
-    void launchDetail(int recordId);
+    void launchDetail(long recordId);
     void deleteRecord(DataRecord targetRecord);
 }
